+++ packages/vendure-plugin-shipping-extensions/src/index.ts
@@ -1,0 +1,3 @@
+export * from './shipping-extensions.plugin';
+export * from './strategies/uk-postalcode-to-geolocation-strategy';
+export * from './strategies/order-address-to-geolocation-strategy';