+++ packages/vendure-plugin-shipping-extensions/src/strategies/order-address-to-geolocation-strategy.ts
@@ -1,0 +1,13 @@
+import { Address } from '@vendure/core';
+import { OrderAddress } from '@vendure/common/lib/generated-shop-types';
+
+export interface OrderAddressToGeolocationConversionStrategy {
+  getGeoLocationForAddress: (
+    orderAddress: OrderAddress
+  ) => Promise<GeoLocation>;
+}
+
+export interface GeoLocation {
+  latitude: number;
+  longitude: number;
+}